:truck: Rename: 고민상담 페이지 폴더 이동
